Move ApiController actions to respective entity controllers and create single action for handling json or html project search results
