+++ src/AppBundle/Controller/MainController.php
@@ -52,39 +52,4 @@ class MainController
             'searchPackageForm' => $searchPackageForm->createView()
         ];
     }
-
-    /**
-     * @Route(
-     *     "/findprojects/{packageSlug}.{id}/{operator}/{versionString}",
-     *     name="searchProjectsThatUsePackage",
-     *     defaults={"operator": "all", "versionString": "1.0.0"},
-     *     requirements={
-     *         "operator": "(==|>=|<=|>|<|all)"
-     *     }
-     * )
-     * @ParamConverter("package", class="AppBundle:Package")
-     * @Template()
-     */
-    public function searchProjectsThatUsePackageAction(Package $package, $operator, $versionString)
-    {
-        $projects = [];
-        $versionConstraint = new VersionConstraint($operator, $versionString);
-
-        foreach($package->getVersions() as $packageVersion) {
-            if($versionConstraint->matches($packageVersion)) {
-                if(count($projects = $packageVersion->getProjects()) === 0) {
-                    continue;
-                }
-                foreach($projects as $project) {
-                    $projects[] = $project;
-                }
-            }
-        }
-
-        return [
-            'package' => $package,
-            'versionConstraint' => $operator . ' ' . $versionString,
-            'projects' => $projects,
-        ];
-    }
 }
