+++ src/AppBundle/Controller/PackageController.php
@@ -6,6 +6,7 @@ use AppBundle\Entity\Package;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
+use Symfony\Component\HttpFoundation\JsonResponse;
 
 class PackageController
 {
@@ -17,5 +18,20 @@ class PackageController
     public function detailAction(Package $package)
     {
         return ['package' => $package];
+    }
+
+    /**
+     * @Route("/api/package/{id}/versions", name="api-package-versions")
+     * @ParamConverter("package", class="AppBundle:Package")
+     * @return JsonResponse
+     */
+    public function apiPackageVersionsAction(Package $package)
+    {
+        $versions = [];
+        foreach($package->getVersions() as $version) {
+            $versions[] = $version->getNormalizedVersion();
+        }
+        $versions = array_values(array_unique($versions));
+        return new JsonResponse($versions);
     }
 }
