+++ src/AppBundle/Controller/ProjectController.php
@@ -2,7 +2,9 @@
 
 namespace AppBundle\Controller;
 
+use AppBundle\Entity\Package;
 use AppBundle\Entity\Project;
+use AppBundle\Entity\VersionConstraint;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
@@ -17,5 +19,29 @@ class ProjectController
     public function detailAction(Project $project)
     {
         return ['project' => $project];
+    }
+
+    /**
+     * @Route(
+     *     "/project/search/{_format}/{packageSlug}.{id}/{operator}/{versionString}",
+     *     defaults={"operator": "all", "versionString": "1.0.0"},
+     *     requirements={
+     *         "operator": "(==|>=|<=|>|<|all)",
+     *          "_format": "json|html"
+     *     },
+     *     name="search-projects-with-package-matching-versionconstraint"
+     * )
+     * @ParamConverter("package", class="AppBundle:Package")
+     * @Template()
+     */
+    public function searchResultsAction(Package $package, $operator, $versionString)
+    {
+        $versionConstraint = new VersionConstraint($operator, $versionString);
+
+        return [
+            'packageVersions' => $package->getMatchingVersionsWithProjects($versionConstraint),
+            'package' => $package,
+            'versionConstraint' => $operator . ' ' . $versionString
+        ];
     }
 }
