+++ src/AppBundle/Resources/views/Project/searchResults.html.twig
@@ -1,0 +1,26 @@
+{% extends '@App/Base/base.html.twig' %}
+
+{% block main %}
+<div class="row">
+    <div class="col-md-6 offset-md-3">
+        <h2>Matching Projects</h2>
+        <p>Using <a href="{{ path('packageDetail', {'slug': package.slug, 'id': package.id}) }}">{{ package.name }}</a> with version constraint {{ versionConstraint }}</p>
+        <ul class="list-group mb-5">
+            {% for version in packageVersions %}
+                <li class="list-group-item">
+                    <strong>Version {{ version.prettyVersion }}</strong><br/>
+                    {% if version.projects|length > 0 %}
+                        Used by:
+                        {% for project in version.projects %}
+                            <a href="{{ path('projectDetail', {'slug': project.slug, 'id': project.id}) }}">
+                                {{ project.name }}
+                            </a>
+                            {% if not loop.last %}, {% endif %}
+                        {% endfor %}
+                    {% endif %}
+                </li>
+            {% endfor %}
+        </ul>
+    </div>
+</div>
+{% endblock%}
