+++ src/AppBundle/Resources/views/Project/searchResults.json.twig
@@ -1,0 +1,21 @@
+{
+    "package": {
+        "id": "{{ package.id }}",
+        "name": "{{ package.name }}",
+        "versions": [
+            {% for packageVersion in packageVersions %}
+                {
+                    "prettyVersion": "{{ packageVersion.prettyVersion }}",
+                    "projects": [
+                        {% for project in packageVersion.projects %}
+                        {
+                            "id": "{{ project.id }}",
+                            "name": "{{ project.name }}"
+                        }{% if not loop.last %},{% endif %}
+                        {% endfor %}
+                    ]
+                }
+            {% endfor %}
+        ]
+    }
+}
